+++ kim/org.integratedmodelling.kim/src/org/integratedmodelling/kim/model/ComputableResource.java
@@ -94,27 +94,31 @@ public class ComputableResource extends KimStatement implements IContextualizabl
 	private IObservable originalObservable;
 	private boolean copy = false;
 	
-	/**
-	 * If not empty, this is the first of a chain (which cannot be hierarchical).
-	 * For now this only happens with URNs.
-	 */
-	private List<ComputableResource> siblings = new ArrayList<>();
+//	/**
+//	 * If not empty, this is the first of a chain (which cannot be hierarchical).
+//	 * For now this only happens with URNs.
+//	 */
+//	private List<ComputableResource> siblings = new ArrayList<>();
 
 	private List<IAnnotation> externalParameters;
 
 	// these set from the outside if the resource is a merge of others
-	private List<String> mergedUrns;
-	private IGeometry mergedGeometry;
-	private IArtifact.Type mergedType;
+	/*
+	 * TODO change to a single merged URN in the local.merged catalog using 
+	 * the ResourcesMergingAdapter, remove all the special handling
+	 */
+//	private List<String> mergedUrns;
+//	private IGeometry mergedGeometry;
+//	private IArtifact.Type mergedType;
 
-	@Override
-	public List<String> getMergedUrns() {
-		return mergedUrns;
-	}
-
-	public void setMergedUrns(List<String> mergedUrns) {
-		this.mergedUrns = mergedUrns;
-	}
+//	@Override
+//	public List<String> getMergedUrns() {
+//		return mergedUrns;
+//	}
+//
+//	public void setMergedUrns(List<String> mergedUrns) {
+//		this.mergedUrns = mergedUrns;
+//	}
 
 	public ComputableResource copy() {
 		ComputableResource ret = new ComputableResource(getEObject(), getParent());
@@ -136,8 +140,8 @@ public class ComputableResource extends KimStatement implements IContextualizabl
 		ret.target = this.target;
 		ret.targetId = this.targetId;
 		ret.copy = true;
-		ret.mergedUrns = this.mergedUrns;
-		ret.mergedType = this.mergedType;
+//		ret.mergedUrns = this.mergedUrns;
+//		ret.mergedType = this.mergedType;
 		ret.interactiveParameters = this.interactiveParameters;
 		ret.externalParameters = this.externalParameters;
 		// ret.type = this.type;
@@ -294,11 +298,11 @@ public class ComputableResource extends KimStatement implements IContextualizabl
 		this.resolutionMode = Mode.RESOLUTION;
 	}
 	
-	public ComputableResource(List<String> mergedUrns, Mode mode, IArtifact.Type type) {
-		this.mergedUrns = mergedUrns;
-		this.mergedType = type;
-		this.resolutionMode = mode;
-	}
+//	public ComputableResource(List<String> mergedUrns, Mode mode, IArtifact.Type type) {
+//		this.mergedUrns = mergedUrns;
+//		this.mergedType = type;
+//		this.resolutionMode = mode;
+//	}
 
 	public ComputableResource(IValueMediator from, IValueMediator to) {
 		this.conversion = new Pair<>(from, to);
@@ -374,11 +378,10 @@ public class ComputableResource extends KimStatement implements IContextualizabl
 			this.literal = Kim.INSTANCE.parseLiteral(value.getLiteral(), Kim.INSTANCE.getNamespace(value));
 			this.type = Type.LITERAL;
 			// this.type = Utils.getArtifactType(this.literal.getClass());
-		} else if (value.getModel() != null) {
-			// only accepted when merging
-			this.urn = value.getModel();
-			this.type = Type.RESOURCE;
-		}
+		} /*
+			 * else if (value.getModel() != null) { // only accepted when merging this.urn =
+			 * value.getModel(); this.type = Type.RESOURCE; }
+			 */
 
 		this.language = value.getLanguage();
 	}
@@ -529,18 +532,18 @@ public class ComputableResource extends KimStatement implements IContextualizabl
 		}
 	}
 
-	/**
-	 * Chain a new resource to the current one. Only happens with URNs so far.
-	 * 
-	 * @param validate
-	 */
-	public void chainResource(ComputableResource resource) {
-		siblings.add(resource);
-	}
-
-	public List<ComputableResource> getSiblings() {
-		return siblings;
-	}
+//	/**
+//	 * Chain a new resource to the current one. Only happens with URNs so far.
+//	 * 
+//	 * @param validate
+//	 */
+//	public void chainResource(ComputableResource resource) {
+//		siblings.add(resource);
+//	}
+//
+//	public List<ComputableResource> getSiblings() {
+//		return siblings;
+//	}
 
 	@Override
 	public void visit(Visitor visitor) {
@@ -554,9 +557,9 @@ public class ComputableResource extends KimStatement implements IContextualizabl
 		if (condition != null) {
 			condition.visit(visitor);
 		}
-		for (ComputableResource sibling : siblings) {
-			sibling.visit(visitor);
-		}
+//		for (ComputableResource sibling : siblings) {
+//			sibling.visit(visitor);
+//		}
 	}
 
 	@Override
@@ -752,9 +755,9 @@ public class ComputableResource extends KimStatement implements IContextualizabl
 				type = Type.CONDITION;
 			} else if (this.urn != null) {
 				type = Type.RESOURCE;
-			} else if (this.mergedUrns != null) {
-				type = Type.MERGED_RESOURCES;
-			}
+			} /*
+				 * else if (this.mergedUrns != null) { type = Type.MERGED_RESOURCES; }
+				 */
 		}
 
 		if (type /* still */ == null) {
@@ -784,9 +787,9 @@ public class ComputableResource extends KimStatement implements IContextualizabl
 	@Override
 	public IGeometry getGeometry() {
 		
-		if (this.mergedGeometry != null) {
-			return this.mergedGeometry;
-		}
+//		if (this.mergedGeometry != null) {
+//			return this.mergedGeometry;
+//		}
 		
 		switch(getType()) {
 		case RESOURCE:
@@ -831,8 +834,8 @@ public class ComputableResource extends KimStatement implements IContextualizabl
 		this.variable = variable;
 	}
 	
-	public void setMergedGeometry(IGeometry geometry) {
-		this.mergedGeometry = geometry;
-	}
+//	public void setMergedGeometry(IGeometry geometry) {
+//		this.mergedGeometry = geometry;
+//	}
 	
 }