+++ klab.engine/src/main/java/org/integratedmodelling/klab/engine/resources/Project.java
@@ -121,9 +121,10 @@ public class Project implements IProject {
 
 		File ret = new File(getRoot() + File.separator + IKimProject.SOURCE_FOLDER + File.separator
 				+ namespaceId.replace('.', File.separatorChar) + ".kim");
-		new File(MiscUtilities.getFilePath(ret.toString())).mkdirs();
+		File npath = new File(MiscUtilities.getFilePath(ret.toString()));
+		npath.mkdirs();
 		try (PrintWriter out = new PrintWriter(ret)) {
-			out.print((createScenario ? "scenario " : (isPrivate ? "private " : "") + "namespace ") + namespaceId
+			out.print((createScenario ? "scenario " : (isPrivate ? "private " : "")) + "namespace " + namespaceId
 					+ ";\n\n");
 		} catch (Exception e) {
 			throw new KlabIOException(e);